+++ game/item.h
@@ -9,7 +9,7 @@ typedef enum {
   ITEM_ANIM_COIN = 0,
 } item_anim_t;
 
-extern int item_count;
+extern int16_t item_count;
 
 void
 item_init(void);
@@ -21,8 +21,6 @@ void
 item_render(frame_buffer_t fb);
 void
 item_update(sprite_t* p);
-//void
-//item_addCoin(uint32_t x, uint32_t y, unsigned short* tilePtr);
 void
-item_add(int x, int y, int anim, unsigned short* tilePtr);
+item_add(int16_t x, int16_t y, int16_t anim, unsigned short* tilePtr);
 #endif
